One more class, fillTableListWithKeyStoreContent() moved to the CertManager.
